fix same-object copies in `deepcopy`
Previously same-object references were not copied correctly and the
following would fail.
```
x = Indices()
y = deepcopy((x, x))
@test y[1] === y[2]
```
In order to not copy the same object multiple times in `deepcopy`, an
`IdDict` is passed to `deepcopy_internal` which should be used by
implementations.
This commit makes use of that dictionary to store already copied objects
and fix the above issue.